fix: implement changes from latest sync
