+++ packages/valory/skills/decision_making_abci/tasks/score_preparations.py
@@ -32,6 +32,7 @@ class ScorePreparation(TaskPreparation):
 
     def check_extra_conditions(self):
         """Check extra conditions"""
+        yield
         return True
 
     def _pre_task(self):
