+++ packages/valory/skills/decision_making_abci/tasks/week_in_olas_preparations.py
@@ -34,6 +34,7 @@ class WeekInOlasCreatePreparation(TaskPreparation):
 
     def check_extra_conditions(self):
         """Check extra conditions"""
+        yield
         return True
 
     def _pre_task(self):
